vuln bridges need gdal, the detection container doesn't
