+++ utils.py
@@ -17,7 +17,6 @@ from pathlib import Path
 from typing import Any, Dict, List, NamedTuple, Tuple, Union
 
 import cv2
-import geojson
 import geopy.distance
 import numpy as np
 import numpy.typing as npt
@@ -26,7 +25,7 @@ import yaml
 from detectron2.data import DatasetCatalog, MetadataCatalog
 from detectron2.data.datasets import load_coco_json, register_coco_instances
 from detectron2.structures import BoxMode
-from osgeo import osr  # pylint: disable-all
+
 from PIL import Image
 from shapely.geometry import LineString, Point
 from shapely.ops import nearest_points
@@ -723,40 +722,6 @@ def get_permit_locations(
             lonlat = [location.latitude, location.longitude]
             permit_locations.append(lonlat)
     return permit_locations
-
-
-def get_bridge_information(file: Union[Path, str]) -> List[List[List[float]]]:
-    """
-    Return a list of coordinates where to find vulnerable bridges and canal walls
-    """
-
-    def rd_to_wgs(coordinates: List[float]) -> List[float]:
-        """
-        Convert rijksdriehoekcoordinates into WGS84 cooridnates. Input parameters: x (float), y (float).
-        """
-        epsg28992 = osr.SpatialReference()
-        epsg28992.ImportFromEPSG(28992)
-
-        epsg4326 = osr.SpatialReference()
-        epsg4326.ImportFromEPSG(4326)
-
-        rd2latlon = osr.CoordinateTransformation(epsg28992, epsg4326)
-        lonlatz = rd2latlon.TransformPoint(coordinates[0], coordinates[1])
-        return [float(value) for value in lonlatz[:2]]
-
-    bridges_coords = []
-    with open(file) as f:
-        gj = geojson.load(f)
-    features = gj["features"]
-    print("Parsing the bridges information")
-    for feature in tqdm(features):
-        bridge_coords = []
-        if feature["geometry"]["coordinates"]:
-            for idx, coords in enumerate(feature["geometry"]["coordinates"][0]):
-                bridge_coords.append(rd_to_wgs(coords))
-            # only add to the list when there are coordinates
-            bridges_coords.append(bridge_coords)
-    return bridges_coords
 
 
 def get_container_locations(file: Path) -> List[List[float]]:
